+++ src/main/java/bridge/exception/ErrorMsg.java
@@ -1,8 +1,8 @@
 package bridge.exception;
 
-import bridge.model.BridgeInfo;
-import bridge.model.Command;
-import bridge.model.Direction;
+import bridge.model.bridge.BridgeInfo;
+import bridge.model.game.Command;
+import bridge.model.game.Direction;
 
 public enum ErrorMsg {
     WRONG_BRIDGE_SIZE("[ERROR] 다리의 크기는 " + BridgeInfo.MIN.getValue() + "이상 " + BridgeInfo.MAX.getValue() + "이하입니다.\n"),
